requirements.txt
-i https://pypi.org/simple
aiohttp==3.8.4 ; python_version >= '3.6'
aiosignal==1.3.1 ; python_version >= '3.7'
ape-solidity @ git+https://github.com/derekpierre/ape-solidity.git@a98042d371bc8e181b52aaa821a738e94a637ab9
appnope==0.1.3 ; sys_platform == 'darwin'
asn1crypto==1.5.1
asttokens==2.2.1
async-timeout==4.0.2 ; python_version >= '3.6'
attrs==23.1.0 ; python_version >= '3.7'
backcall==0.2.0
base58==1.0.3
bitarray==2.7.3
black==23.3.0
cached-property==1.5.2
cachetools==5.3.0 ; python_version ~= '3.7'
certifi==2023.5.7 ; python_version >= '3.6'
cffi==1.15.1
cfgv==3.3.1 ; python_full_version >= '3.6.1'
chardet==5.1.0 ; python_version >= '3.7'
charset-normalizer==3.1.0 ; python_full_version >= '3.7.0'
click==8.1.3 ; python_version >= '3.7'
coincurve==18.0.0
colorama==0.4.6 ; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4, 3.5, 3.6'
commonmark==0.9.1
cryptography==40.0.2
cytoolz==0.12.1 ; implementation_name == 'cpython'
dataclassy==0.11.1 ; python_version >= '3.6'
decorator==5.1.1 ; python_version >= '3.5'
deprecated==1.2.13 ; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
distlib==0.3.6
eip712==0.2.1 ; python_version >= '3.8' and python_version < '4'
eth-abi==4.0.0 ; python_version >= '3.7' and python_version < '4'
eth-account==0.8.0 ; python_version >= '3.6' and python_version < '4'
eth-ape==0.6.9
eth-bloom==2.0.0 ; python_version >= '3.7' and python_version < '4'
eth-hash[pycryptodome]==0.5.1 ; python_version >= '3.7' and python_version < '4'
eth-keyfile==0.6.1
eth-keys==0.4.0
eth-rlp==0.3.0 ; python_version >= '3.7' and python_version < '4'
eth-tester[py-evm]==0.9.0b1
eth-typing==3.3.0 ; python_full_version >= '3.7.2' and python_version < '4'
eth-utils==2.1.0 ; python_version >= '3.7' and python_version < '4'
ethpm-types==0.5.0 ; python_version >= '3.8' and python_version < '4'
evm-trace==0.1.0a20 ; python_version >= '3.8' and python_version < '4'
exceptiongroup==1.1.1 ; python_version < '3.11'
executing==1.2.0
filelock==3.12.0 ; python_version >= '3.7'
flake8==6.0.0
frozenlist==1.3.3 ; python_version >= '3.7'
hexbytes==0.3.0 ; python_version >= '3.7' and python_version < '4'
identify==2.5.24 ; python_version >= '3.7'
idna==3.4 ; python_version >= '3.5'
ijson==3.2.0.post0
importlib-metadata==6.6.0 ; python_version >= '3.7'
importlib-resources==5.12.0 ; python_version < '3.9'
iniconfig==2.0.0 ; python_version >= '3.7'
ipython==8.12.2 ; python_version >= '3.8'
isort==6.0.0b2
jedi==0.18.2 ; python_version >= '3.6'
jsonschema==4.18.0a7 ; python_version >= '3.8'
jsonschema-specifications==2023.5.1 ; python_version >= '3.8'
lazyasd==0.1.4
lru-dict==1.1.8
matplotlib-inline==0.1.6 ; python_version >= '3.5'
mccabe==0.7.0 ; python_version >= '3.6'
morphys==1.0
msgspec==0.15.0 ; python_version >= '3.8'
multidict==6.0.4 ; python_version >= '3.7'
mypy-extensions==0.4.4 ; python_version >= '2.7'
nodeenv==1.8.0 ; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4, 3.5, 3.6'
nucypher-core==0.7.0
numpy==1.24.3 ; python_version < '3.10'
packaging==23.1 ; python_version >= '3.7'
pandas==1.5.3 ; python_version >= '3.8'
parsimonious==0.9.0
parso==0.8.3 ; python_version >= '3.6'
pathspec==0.11.1 ; python_version >= '3.7'
pexpect==4.8.0 ; sys_platform != 'win32'
pickleshare==0.7.5
pkgutil-resolve-name==1.3.10 ; python_version < '3.9'
platformdirs==3.5.1 ; python_version >= '3.7'
pluggy==1.0.0 ; python_version >= '3.6'
pre-commit==3.3.2
prompt-toolkit==3.0.38 ; python_full_version >= '3.7.0'
protobuf==4.23.1 ; python_version >= '3.7'
ptyprocess==0.7.0
pure-eval==0.2.2
py-cid==0.3.0
py-ecc==6.0.0 ; python_version >= '3.6' and python_version < '4'
py-evm==0.7.0a2
py-geth==3.12.0 ; python_version >= '3'
py-multibase==1.0.3
py-multicodec==0.2.1
py-multihash==0.2.3
py-solc-x==1.1.1 ; python_version >= '3.6' and python_version < '4'
pycodestyle==2.10.0 ; python_version >= '3.6'
pycparser==2.21
pycryptodome==3.17
pydantic==1.10.7 ; python_version >= '3.7'
pyethash==0.1.27
pyflakes==3.0.1 ; python_version >= '3.6'
pygithub==1.58.2 ; python_version >= '3.7'
pygments==2.15.1 ; python_version >= '3.7'
pyjwt[crypto]==2.7.0 ; python_version >= '3.7'
pynacl==1.5.0 ; python_version >= '3.6'
pyproject-api==1.5.1 ; python_version >= '3.7'
pysha3==1.0.2
pytest==7.3.1 ; python_version >= '3.7'
python-baseconv==1.2.2
python-dateutil==2.8.2 ; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2'
pytz==2023.3
pyyaml==6.0 ; python_version >= '3.6'
referencing==0.28.2 ; python_version >= '3.8'
regex==2023.5.5 ; python_version >= '3.6'
requests==2.30.0 ; python_version >= '3.7'
rich==12.6.0 ; python_full_version >= '3.6.3' and python_full_version < '4.0.0'
rlp==3.0.0
rpds-py==0.7.1 ; python_version >= '3.8'
semantic-version==2.10.0 ; python_version >= '2.7'
setuptools==67.7.2 ; python_version >= '3.7'
six==1.16.0 ; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2'
sortedcontainers==2.4.0
sqlalchemy==2.0.14 ; python_version >= '3.7'
stack-data==0.6.2
tomli==2.0.1 ; python_version < '3.11'
toolz==0.12.0 ; python_version >= '3.5'
tox==4.5.1
tqdm==4.65.0 ; python_version >= '3.7'
traitlets==5.9.0 ; python_version >= '3.7'
trie==2.1.0 ; python_version >= '3.7' and python_version < '4'
typing-extensions==4.5.0 ; python_version < '3.10'
urllib3==2.0.2 ; python_version >= '3.7'
varint==1.0.2
virtualenv==20.23.0 ; python_version >= '3.7'
watchdog==2.3.1 ; python_version >= '3.6'
wcwidth==0.2.6
web3[tester]==6.4.0 ; python_full_version >= '3.7.2'
websockets==11.0.3 ; python_version >= '3.7'
wrapt==1.15.0 ; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4'
yarl==1.9.2 ; python_version >= '3.7'
zipp==3.15.0 ; python_version >= '3.7'
